+++ internal/connector/service_daemon/service_daemon_openbsd.go
@@ -1,7 +1,7 @@
 //go:build openbsd
 // +build openbsd
 
-package service
+package service_daemon
 
 import "fmt"
